Creating a generic ResultViewModel

+++ Blog/Controllers/CategoryController.cs
@@ -11,14 +11,17 @@ namespace Blog.Controllers
         [HttpGet("v1/categories")]
         public async Task<IActionResult> GetAsync([FromServices] BlogDataContext context)
         {
+            if (!ModelState.IsValid)
+                return BadRequest();
+
             try
             {
                 var categories = await context.Categories.ToListAsync();
-                return Ok(categories);
+                return Ok(new ResultViewModel<List<Category>>(categories));
             }
-            catch (Exception)
+            catch
             {
-                return StatusCode(500, "05X04 - Falha interna no servidor");
+                return StatusCode(500, new ResultViewModel<List<Category>>("05X04 - Falha interna no servidor"));
             }
         }
 
@@ -26,18 +29,21 @@ namespace Blog.Controllers
         public async Task<IActionResult> GetByIdAsync([FromServices] BlogDataContext context,
                                                       [FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest();
+
             try
             {
                 var category = await context.Categories.FirstOrDefaultAsync(x => x.Id == id);
 
                 if (category == null)
-                    return NotFound();
+                    return NotFound(new ResultViewModel<Category>("Não foi possivel encontrar a categoria"));
 
-                return Ok(category);
+                return Ok(new ResultViewModel<Category>(category));
             }
             catch (Exception)
             {
-                return StatusCode(500, "05X05 - Falha interna no servidor");
+                return StatusCode(500, new ResultViewModel<Category>("05X05 - Falha interna no servidor"));
             }
         }
 
@@ -45,6 +51,9 @@ namespace Blog.Controllers
         public async Task<IActionResult> PostAsync([FromServices] BlogDataContext context,
                                                    [FromBody] EditorCategoryViewModel model)
         {
+            if (!ModelState.IsValid)
+                return BadRequest();
+
             try
             {
                 var category = new Category()
@@ -56,15 +65,15 @@ namespace Blog.Controllers
 
                 await context.Categories.AddAsync(category);
                 await context.SaveChangesAsync();
-                return Created($"v1/categories/{category.Id}", category);
+                return Created($"v1/categories/{category.Id}", new ResultViewModel<Category>(category));
             }
             catch (DbUpdateException)
             {
-                return StatusCode(500, "05XE9 - Não foi possivel incluir a categoria");
+                return StatusCode(500, new ResultViewModel<Category>("05XE9 - Não foi possivel incluir a categoria"));
             }
             catch (Exception)
             {
-                return StatusCode(500, "05XE10 - Falha interna no servidor");
+                return StatusCode(500, new ResultViewModel<Category>("05XE10 - Falha interna no servidor"));
             }
         }
 
@@ -73,12 +82,15 @@ namespace Blog.Controllers
                                                   [FromRoute] int id,
                                                   [FromBody] EditorCategoryViewModel model)
         {
+            if (!ModelState.IsValid)
+                return BadRequest();
+
             try
             {
                 var category = await context.Categories.FirstOrDefaultAsync(x => x.Id == id);
 
                 if (category == null)
-                    return NotFound();
+                    return NotFound(new ResultViewModel<Category>("Não foi possivel encontrar a categoria"));
 
                 category.Name = model.Name;
                 category.Slug = model.Slug;
@@ -90,11 +102,11 @@ namespace Blog.Controllers
             }
             catch (DbUpdateException)
             {
-                return StatusCode(500, "05XE8 - Não foi possivel alterar a categoria");
+                return StatusCode(500, new ResultViewModel<Category>("05XE8 - Não foi possivel alterar a categoria"));
             }
             catch (Exception)
             {
-                return StatusCode(500, "05XE11 - Falha interna no servidor");
+                return StatusCode(500, new ResultViewModel<Category>("05XE11 - Falha interna no servidor"));
             }
         }
 
@@ -102,12 +114,15 @@ namespace Blog.Controllers
         public async Task<IActionResult> DeleteAsync([FromServices] BlogDataContext context,
                                                      [FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest();
+
             try
             {
                 var category = await context.Categories.FirstOrDefaultAsync(x => x.Id == id);
 
                 if (category == null)
-                    return NotFound();
+                    return NotFound(new ResultViewModel<Category>("Não foi possivel encontrar a categoria"));
 
                 context.Categories.Remove(category);
                 await context.SaveChangesAsync();
@@ -116,11 +131,11 @@ namespace Blog.Controllers
             }
             catch (DbUpdateException)
             {
-                return StatusCode(500, "05XE7 - Não foi possivel excluir a categoria");
+                return StatusCode(500, new ResultViewModel<Category>("05XE7 - Não foi possivel excluir a categoria"));
             }
             catch (Exception)
             {
-                return StatusCode(500, "05XE12 - Falha interna no servidor");
+                return StatusCode(500, new ResultViewModel<Category>("05XE12 - Falha interna no servidor"));
             }
         }
     }
